requirements.txt
functions-framework==3.*
discord.py
dotenv
flask